Added code to method.m to allow real time measurement of laser fluctuations. There is a new field in the object generated -- obj.laserPD, which contains the output of the external photodiode capturing laser scattering.
Added a function to subtract a background (measured at the start of the experiment) from the photodiode output. Currently, it seems we have to re-background every time we switch a method, so we need to fix that. Probably something related to the BackgroundReset function in method.m.

Needs to be tested with hardware.

Removed one line that had previously been added to method_show_spectrum.m, which was no longer needed.