Fixed how it was starting and closing
It didn't know what wavelength it was at when it started. The iHR's
activeX controller only returns the correct wavelength when the
specrograph starts at turret 0. (We don't know why.) So on closing the
Spectrometer.m, we now 1) save current turret, 2) the move to turret 0
3) finish closing. Then on startup, initialize in turret 0, then read
the old turret and move back there. Seems to work...