Fixed issues plotting the third-order response and changed initial params
During the operation of 'Method_2d_quasiphase.m' the third-order response is supposed to be plotted in the main plot at the end of each scan. This was not happening. The plot was occuring on a secondary plot outside of the GUI. In order to fix this the 'hMainAxes' was assigned as the parent of the contour plot. Then in 'Method.m' in the 'RefreshPlots' method, if the plot is detected to be a contour plot it resets the level list to be appropriate for the displayed data.

In 'Method_2d_quasiphase.m', the initial parameters were also altered to reflect more frequently used parameters (nscans: 10 -> 2; speed: 1700 -> 800).